+++ hi_streaming/hi_streaming/MonolithAudioFormat.h
@@ -406,7 +406,7 @@ public:
 
 	int64 getMonolithLength(int sampleIndex) const
 	{
-		return jmax<int64>(0, multiChannelSampleInformation[0][sampleIndex].length);
+		return (int64)jmax<int>(0, (int)multiChannelSampleInformation[0][sampleIndex].length);
 	}
 
 	double getMonolithSampleRate(int sampleIndex) const
